chore: restructure the project
