--- src/components/introduction/format/TextLink.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-interface TextLinkProps {
-    link: string;
-    children: React.ReactNode;
-}
-export function TextLink({ link, children }: TextLinkProps) {
-    return (
-        <a
-            href={link}>{children}</a>
-    )
-}
