--- src/components/introduction/format/mark/TextMark.tsx
@@ -1,7 +1,0 @@
-import './mark.css'
-interface MarkProps {
-    children: React.ReactNode;
-}
-export function TextMark({ children }: MarkProps) {
-    return <mark>{children}</mark>
-}
