--- src/components/introduction/format/mark/mark.css
@@ -1,6 +1,0 @@
-mark {
-    background-color: #444444;
-    color: rgba(255, 255, 255, 0.75);
-    padding: 0.1rem 0.3rem;
-    border-radius: 0.5rem;
-}
