--- src/components/introduction/profile.css
@@ -1,20 +1,0 @@
-.introduction-name {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 0.25rem;
-}
-.introduction-name > img {
-    width: 6rem;
-    border-radius: 1rem;
-}
-.introduction-name > h1 {
-    margin: 0;
-}
-.introduction-name > span {
-    color: rgba(255, 255, 255, 0.75)
-}
-.introduction-description {
-    text-align: center;
-}
