+++ src/hw2/main.tsx
@@ -1,0 +1,5 @@
+export function Homework2() {
+    return (
+        <></>
+    )
+}
